fix: src/__tests__ => src/,tests/
